Added Admin Pages

+++ ETrade/Entities/Concrete/District.cs
@@ -8,7 +8,7 @@ namespace Entities.Concrete
     public class District : IEntity
     {
         public byte CityId { get; set; }
-        public byte DistrictId { get; set; }
+        public short DistrictId { get; set; }
         public string DistrictName { get; set; }
     }
 }
