+++ ETrade/Entities/Concrete/UserRole.cs
@@ -8,6 +8,6 @@ namespace Entities.Concrete
     public class UserRole : IEntity
     {
         public int UserId { get; set; }
-        public int RoleId { get; set; }
+        public byte RoleId { get; set; }
     }
 }
